add error handling and status codes

+++ RealEstateAPi/Controllers/CategoriesController.cs
@@ -13,45 +13,66 @@ namespace RealEstateApi.Controllers
 
         // GET: api/<CategoriesController>
         [HttpGet]
-        public IEnumerable<Category> Get()
+        public IActionResult Get()
         {
-            return _dbContext.Categories;
+            return Ok(_dbContext.Categories);
         }
 
         // GET api/<CategoriesController>/5
         [HttpGet("{id}")]
-        public Category Get(int id)
+        public IActionResult Get(int id)
         {
             var category = _dbContext.Categories.FirstOrDefault(x => x.Id == id);
-            return category;
+            if (category == null)
+            {
+                return NotFound();
+            }
+            return Ok(category);
         }
 
         // POST api/<CategoriesController>
         [HttpPost]
-        public void Post([FromBody] Category category)
+        public IActionResult Post([FromBody] Category category)
         {
             _dbContext.Categories.Add(category);
             _dbContext.SaveChanges();
+            return StatusCode(StatusCodes.Status201Created);
 
         }
 
         // PUT api/<CategoriesController>/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] Category categoryObj)
+        public IActionResult Put(int id, [FromBody] Category categoryObj)
         {
             var category = _dbContext.Categories.Find(id);
-            category.Name = categoryObj.Name;
-            category.ImageUrl = categoryObj.ImageUrl;
-            _dbContext.SaveChanges();
+            if (category == null)
+            {
+                return NotFound("No record found against this id" + id);
+            }
+            else
+            {
+                category.Name = categoryObj.Name;
+                category.ImageUrl = categoryObj.ImageUrl;
+                _dbContext.SaveChanges();
+                return Ok("Record updated succesfully");
+            }
         }
 
         // DELETE api/<CategoriesController>/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
             var category = _dbContext.Categories.Find(id);
-            _dbContext.Categories.Remove(category);
-            _dbContext.SaveChanges();
+            if (category == null)
+            {
+                return NotFound("No record found against this Id: " + id);
+            }
+            else
+            {
+                _dbContext.Categories.Remove(category);
+                _dbContext.SaveChanges();
+                return Ok("Record Deleted");
+            }
         }
     }
 }
